Modules/Web_check.py
import requests
from bs4 import BeautifulSoup
from datetime import datetime, timedelta
import time
import nltk
from nltk.tokenize import sent_tokenize, word_tokenize
from nltk.corpus import stopwords
from nltk.probability import FreqDist
from nltk.tag import pos_tag
from nltk.chunk import ne_chunk

# Download NLTK resources (uncomment the following line if not already downloaded)
nltk.download('punkt')
nltk.download('stopwords')
nltk.download('averaged_perceptron_tagger')
nltk.download('maxent_ne_chunker')
nltk.download('words')

# URL of the webpage to check
url = "https://example.com"

# File to save the summary
summary_file = "summary.txt"

# Function to get the content of the webpage
def get_webpage_content(url):
    response = requests.get(url)
    if response.status_code == 200:
        return response.text
    else:
        return None

# Function to extract new content from the webpage
def extract_new_content(old_content, new_content):
    # Define a function to extract text from HTML content
    def extract_text_from_html(html_content):
        # Placeholder logic to extract text from HTML content
        # You need to implement this function based on your requirements
        return "Extracted text from HTML"

    # Extract text from old and new content
    old_text = extract_text_from_html(old_content)
    new_text = extract_text_from_html(new_content)

    # Compare the extracted text to determine if new content is different
    return old_text != new_text


    # Compare the lengths of the extracted text
    if len(new_text) > len(old_text):
        # If the new text is longer, consider it new content
        return True

    # Compare the number of unique words in the new text
    old_words = set(old_text.split())
    new_words = set(new_text.split())
    if len(new_words - old_words) > 0:
        # If there are new unique words, consider it new content
        return True

    # Implement additional checks based on specific criteria or patterns in the content
    # For example, checking for the presence of specific keywords or changes in structure

    # If no specific criteria indicate new content, return False
    return False

# Function to summarize the new content
def summarize_content(content):
    # Tokenize the content into sentences
    sentences = sent_tokenize(content)

    # Tokenize each sentence into words and remove stopwords
    stop_words = set(stopwords.words('english'))
    word_tokens = [word_tokenize(sentence.lower()) for sentence in sentences]
    filtered_words = [[word for word in words if word not in stop_words] for words in word_tokens]

    # Calculate the frequency distribution of words
    word_freq = FreqDist(word for words in filtered_words for word in words)

    # Get the top 10 most frequent words
    top_words = word_freq.most_common(10)

    # Extract named entities using part-of-speech tagging and chunking
    named_entities = []
    for sentence in sentences:
        words = word_tokenize(sentence)
        tagged_words = pos_tag(words)
        chunked_words = ne_chunk(tagged_words)
        for subtree in chunked_words:
            if type(subtree) == nltk.tree.Tree:
                if subtree.label() == 'NE':
                    named_entities.append(' '.join([word for word, pos in subtree.leaves()]))

    # Create a summary combining top words and named entities
    summary = 'Top words: ' + ', '.join(word for word, freq in top_words) + '. '
    summary += 'Named entities: ' + ', '.join(named_entities) + '.'

    return summary

# Function to save the summary to a file
def save_summary(summary):
    with open(summary_file, "a") as f:
        f.write(summary + "\n\n")

# Function to run the script
def run_script():
    # Get the current time
    current_time = datetime.now()

    # Get the content of the webpage
    new_content = get_webpage_content(url)

    # Load the old content from the summary file
    try:
        with open(summary_file, "r") as f:
            old_content = f.read()
    except FileNotFoundError:
        old_content = None

    # Check if there is new content
    if new_content and extract_new_content(old_content, new_content):
        # Summarize the new content
        summary = summarize_content(new_content)

        # Save the summary to the file
        save_summary(summary)

        print("New content found and summarized.")
    else:
        print("No new content found.")

    # Wait for an hour before running the script again
    time.sleep(3600)

# Run the script continuously
while True:
    run_script()
