requirements.txt
numpy
pandas
matplotlib
bs4
discord.py
rasa
nltk